requirements.txt
alembic==0.9.5
Baker==1.3
certifi==2017.7.27.1
chardet==3.0.4
cheroot==5.1.0
CherryPy==10.1.0
clay-flask==2.1.5
click==6.7
colorama==0.3.9
Django==1.11.4
django-filter==1.0.4
django-haystack==2.6.1
djangorestframework==3.6.3
Flask==0.12.2
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.7
Markdown==2.6.8
MarkupSafe==1.0
Moar==3.0.0
oauthlib==2.0.2
olefile==0.44
Pillow==4.2.1
pkg-resources==0.0.0
portend==2.1.2
Pygments==2.2.0
PyJWT==1.5.2
PyMySQL==0.7.11
python-dateutil==2.6.1
python-editor==1.0.3
python-openid==2.2.5
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
six==1.10.0
social-auth-app-django==1.2.0
social-auth-core==1.4.0
SQLAlchemy==1.1.13
tempora==1.8
urllib3==1.22
Voodoo==1.4.2
Wand==0.4.4
Werkzeug==0.12.2
